refactor: move PDF byte conversion functions to pdf_page_tools and simplify logic
